interactive-data-dashboard/requirements-dev.txt
# Include all production dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
pytest-mock>=3.11.1
pytest-timeout>=2.1.0
pytest-asyncio>=0.21.1

# Code quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0
pre-commit>=3.4.0
isort>=5.12.0

# Security
safety>=2.3.0
bandit>=1.7.5

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0

# Development tools
ipython>=8.15.0
jupyter>=1.0.0
nb-clean>=2.4.0

# Performance profiling
memory-profiler>=0.61.0
line-profiler>=4.1.1

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Database (for advanced features)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
redis>=4.6.0

# File storage (for cloud deployment)
boto3>=1.28.0
minio>=7.1.0

# Additional utilities
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.5.0
tqdm>=4.66.0